[Refactor] 좋아요한메뉴cell 리팩토링
